requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
amqp==5.2.0
aniso8601==9.0.1
appdirs==1.4.4
asgiref==3.8.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
billiard==4.2.0
blinker==1.6.2
build==1.2.1
buildozer==1.5.0
CacheControl==0.14.0
cachetools==5.3.1
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.4
Command==0.1.0
contourpy==1.2.1
crashtest==0.4.1
cron-descriptor==1.4.3
cycler==0.12.1
Cython==3.0.10
distlib==0.3.8
dj-database-url==2.2.0
Django==4.2
django-celery-beat==2.6.0
django-filter==24.2
django-timezone-field==6.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
docutils==0.21.2
drf-yasg==1.21.7
dulwich==0.21.7
et-xmlfile==1.1.0
exceptiongroup==1.1.0
fastjsonschema==2.19.1
ffmpeg-python==0.2.0
filelock==3.13.4
fonttools==4.53.0
frozenlist==1.3.3
fsspec==2024.6.0
future==0.18.3
gitdb==4.0.11
GitPython==3.1.43
gmpy2==2.1.5
google-api-core==2.12.0
google-api-python-client==2.101.0
google-auth==2.23.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.60.0
greenlet==1.1.2
gspread==6.1.2
gunicorn==22.0.0
httplib2==0.22.0
idna==3.7
importlib_metadata==7.1.0
inflection==0.5.1
iniconfig==2.0.0
installer==0.7.0
isodate==0.6.1
itsdangerous==2.1.2
jaraco.classes==3.4.0
Jinja2==3.1.2
jsonify==0.5
jsonschema==4.1.0
keyring==24.3.1
Kivy==2.3.0
Kivy-Garden==0.1.5
kiwisolver==1.4.5
kkconst==1.1.3
kombu==5.3.7
llvmlite==0.39.1
magic-filter==1.0.9
MarkupSafe==2.1.2
marshmallow==3.19.0
matplotlib==3.9.0
more-itertools==9.1.0
mpmath==1.2.1
msgpack==1.0.8
multidict==6.0.4
networkx==3.0
numba==0.56.4
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
openai==0.27.2
openai-whisper @ git+https://github.com/openai/whisper.git@b5851c6c40e753606765ac45b85b298e3ae9e00d
opencv-python==4.10.0.82
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pexpect==4.9.0
pillow==10.3.0
pkginfo==1.10.0
platformdirs==4.2.1
pluggy==1.0.0
poetry==1.8.2
poetry-core==1.9.0
poetry-plugin-export==1.7.1
prompt-toolkit==3.0.43
protobuf==4.24.3
psutil==5.9.8
psycopg2-binary==2.9.9
ptyprocess==0.7.0
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
pyparsing==3.1.1
pyproject_hooks==1.0.0
pyrsistent==0.18.0
pyTelegramBotAPI==4.10.0
pytest==7.2.1
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.8.1
regex==2023.3.23
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9
scipy==1.13.1
seaborn==0.13.2
sh==1.14.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.15
sqlparse==0.5.0
StatusCode==1.0.0
stripe==9.5.0
sympy==1.11.1
tiktoken==0.3.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.4
torch==2.3.0
torchvision==0.18.0
tqdm==4.65.0
trove-classifiers==2024.4.10
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.2.28
ultralytics-thop==0.2.7
uritemplate==4.1.1
urllib3==1.26.18
vine==5.1.0
virtualenv==20.26.0
wcwidth==0.2.13
Werkzeug==2.3.4
xattr==1.1.0
yarl==1.9.2
zipp==3.6.0
